+++ History.md
@@ -3,6 +3,7 @@ Unpublished
   * add coin acceptor stress test
   * replace new Buffer with Buffer.from or Buffer.alloc
   * sleep for one second after exporting gpios to avoid permission issues
+  * construct AsyncResource for callback when callback is stored
 
 2.0.1 - Apr 02 2018
 ===================